+++ example/queries.dart
@@ -69,7 +69,7 @@ void main() async {
   await coll
       .find(where.jsQuery('this.my_field % 100 == 35'))
       .forEach((v) => print(v));
-  var count = await coll.legacyCount(where.gt('my_field', 995));
+  var count = await coll.count(where.gt('my_field', 995));
   print('Count of records with my_field > 995: $count');
   var databases = await db.listDatabases();
   print('List of databases: $databases');